+++ diesel/src/pg/expression/operators.rs
@@ -8,5 +8,7 @@ infix_operator!(Contains, " @> ", backend: Pg);
 infix_operator!(IsContainedBy, " <@ ", backend: Pg);
 infix_operator!(ILike, " ILIKE ", backend: Pg);
 infix_operator!(NotILike, " NOT ILIKE ", backend: Pg);
+infix_operator!(SimilarTo, " SIMILAR TO ", backend: Pg);
+infix_operator!(NotSimilarTo, " NOT SIMILAR TO ", backend: Pg);
 postfix_operator!(NullsFirst, " NULLS FIRST", NotSelectable, backend: Pg);
 postfix_operator!(NullsLast, " NULLS LAST", NotSelectable, backend: Pg);